MPN-1004: Ex. 8.9
Implement and test Prop.and and Prop.or.

Use new implementations of Result and Prop as suggested in the book.

Rename Prop.&& and Prop.|| to Prop.and and Prop.or respectively since
we're not using them as operators any more but ratther as regular
methods now.

Reimplement the tests for the above two methods to reflect the new
implementations of Prop and Result.